Correctly import bazaar when testing.
When running tests with `python bazaar/test/test.py` the statement
`import bazaar` would raise an ImportError because it wasn't able
to find the package, which is its parent directory.
This fix tries to import it normally but when it case it fails,
it adds the parent directory to the system path and retries.