Let the Buffer be used as a context handler.
Previously, FileSystem.open method wouldn't work in a with statement
since it was returning the BufferWrapper object which didn't implement
the __enter__ and __exit__ methods.

Even though it would seem like the __getattr__ method should catch
the calls to these attributes and redirect them to the wrapped
object, due to the way Python itself works, it won't.
Both methods for context handlers must be manually stated in the
class.

Related info: https://stackoverflow.com/questions/4146095/