Correct list of fields from a file to be fetched from Mongo
The projection was formed as {field: -1}. Due to the way pymongo
works, by including each key which has a "truthy" value on the
dictionary, it was working as bool(-1) is True.

However, the negative one was confusing and probably unintentional.
Just change it to True for clarity and conciseness.

Related PyMongo documentation: https://pymongo.readthedocs.io/en/stable/api/pymongo/collection.html#pymongo.collection.Collection.find